Order the code to prevent depandncy issues
